+++ tests/e2e/authz/client/testutil/cli_test.go
@@ -1,0 +1,20 @@
+//go:build norace
+// +build norace
+
+package testutil
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/suite"
+
+	"github.com/cosmos/cosmos-sdk/simapp"
+	"github.com/cosmos/cosmos-sdk/testutil/network"
+	"github.com/cosmos/cosmos-sdk/x/authz/client/testutil"
+)
+
+func TestIntegrationTestSuite(t *testing.T) {
+	cfg := network.DefaultConfig(simapp.NewTestNetworkFixture)
+	cfg.NumValidators = 1
+	suite.Run(t, testutil.NewIntegrationTestSuite(cfg))
+}
