+++ tests/e2e/crisis/client/testsuite/cli_test.go
@@ -1,0 +1,20 @@
+//go:build norace
+// +build norace
+
+package testsuite
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/suite"
+
+	"github.com/cosmos/cosmos-sdk/simapp"
+	"github.com/cosmos/cosmos-sdk/testutil/network"
+	"github.com/cosmos/cosmos-sdk/x/crisis/client/testsuite"
+)
+
+func TestIntegrationTestSuite(t *testing.T) {
+	cfg := network.DefaultConfig(simapp.NewTestNetworkFixture)
+	cfg.NumValidators = 1
+	suite.Run(t, testutil.NewIntegrationTestSuite(cfg))
+}
