+++ tests/e2e/distribution/client/testutil/cli_test.go
@@ -1,0 +1,21 @@
+package testutil
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/suite"
+
+	"github.com/cosmos/cosmos-sdk/x/distribution/client/testutil"
+)
+
+func TestIntegrationTestSuite(t *testing.T) {
+	suite.Run(t, new(testutil.IntegrationTestSuite))
+}
+
+func TestGRPCQueryTestSuite(t *testing.T) {
+	suite.Run(t, new(testutil.GRPCQueryTestSuite))
+}
+
+func TestWithdrawAllSuite(t *testing.T) {
+	suite.Run(t, new(testutil.WithdrawAllTestSuite))
+}
