--- x/distribution/client/testutil/cli_test.go
@@ -1,19 +1,0 @@
-package testutil
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/suite"
-)
-
-func TestIntegrationTestSuite(t *testing.T) {
-	suite.Run(t, new(IntegrationTestSuite))
-}
-
-func TestGRPCQueryTestSuite(t *testing.T) {
-	suite.Run(t, new(GRPCQueryTestSuite))
-}
-
-func TestWithdrawAllSuite(t *testing.T) {
-	suite.Run(t, new(WithdrawAllTestSuite))
-}
